backend/requirements.txt
anyio==4.4.0
click==8.1.7
exceptiongroup==1.2.2
h11==0.14.0
idna==3.7
psycopg2-binary==2.9.9
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.5